binary conversion

+++ _putnum.c
@@ -2,15 +2,14 @@
 #include "main.h"
 
 /**
- * _putnum_withCounter - writes the character c to stdout
- * @num: The character to print
+ * _putnum_withCounter - writes the number to stdout
+ * @num: The number to print
  * @i: pointer to length
  *
  * Return: On success length of number.
- * On error, -1 is returned, and errno is set appropriately.
  */
 
-void _putnum_withCounter(long int num, int *i)
+void _putnum_withCounter(int num, int *i)
 {
 	if (num / 10)
 		_putnum_withCounter(num / 10, i);
@@ -19,13 +18,12 @@ void _putnum_withCounter(long int num, int *i)
 }
 
 /**
- * _putnum - writes the character c to stdout
- * @num: The character to print
+ * _putnum - writes the number to stdout
+ * @num: The number to print
  *
  * Return: On success length of number.
- * On error, -1 is returned, and errno is set appropriately.
  */
-int _putnum(long int num)
+int _putnum(int num)
 {
 	int i;
